+++ build_effective_set_generator_java/Dockerfile
@@ -21,7 +21,7 @@ COPY --from=builder --chown=1001:root /deployments /deployments
 COPY --from=builder /etc/alternatives/jre/lib/security/java.security /etc/alternatives/jre/lib/security/java.security
 
 ## NOTE: This script requires Python and will fail unless Python is added in the future.
-COPY scripts/utils /module/scripts/scripts/utils/
+COPY scripts/utils /module/scripts/utils
 
 RUN chmod g+rwX /deployments
