+++ build_pipegene/scripts/github_actions.py
@@ -1,8 +1,7 @@
 import click
 
 from validations import validate_pipeline
-from scripts.utils.pipeline_parameters import PipelineParametersHandler
-from scripts.utils.log_pipe_params import log_pipeline_params
+from scripts.utils.pipeline_parameters import PipelineParametersHandler, log_pipeline_params
 
 @click.group()
 def cli():
