+++ scripts/utils/pipeline_parameters.py
@@ -1,4 +1,6 @@
+import json
 from os import getenv
+from envgenehelper import logger
 from envgenehelper.plugin_engine import PluginEngine
 
 
@@ -29,6 +31,25 @@ def get_pipeline_parameters() -> dict:
         'DEPLOYMENT_SESSION_ID': getenv("DEPLOYMENT_SESSION_ID", ""),
         'LOG_LEVEL': getenv("LOG_LEVEL")
     }
+
+
+def log_pipeline_params(params: dict):
+   params_str = "Input parameters are: "
+   
+   if params.get("CRED_ROTATION_PAYLOAD"): 
+       params["CRED_ROTATION_PAYLOAD"] = "***"
+
+   for k, v in params.items():
+        try:
+            parsed = json.loads(v)
+            params[k] = json.dumps(parsed, separators=(",", ":"))
+        except (TypeError, ValueError):
+            pass
+
+        params_str += f"\n{k.upper()}: {params[k]}"
+
+   logger.info(params_str)
+
 
 class PipelineParametersHandler:
     def __init__(self, **kwargs):
